Separate store implementation.
