+++ src/lib/locationState/stores/storage/storage-store.ts
@@ -1,12 +1,55 @@
-import { Store } from "../store";
+import { Listener, Store } from "../store";
 
 export const locationKeyPrefix = "__location_state_";
 
-export class StorageStore extends Store {
+export class StorageStore implements Store {
   private currentKey: string | null = null;
+  listeners: Map<string, Set<Listener>> = new Map();
+  state: Record<string, unknown> = {};
 
-  constructor(private readonly storage?: Storage) {
-    super();
+  constructor(private readonly storage?: Storage) {}
+
+  subscribe(name: string, listener: () => void) {
+    const listeners = this.listeners.get(name);
+    if (listeners) {
+      listeners.add(listener);
+    } else {
+      this.listeners.set(name, new Set([listener]));
+    }
+    return () => this.unsubscribe(name, listener);
+  }
+
+  private unsubscribe(name: string, listener: Listener) {
+    const listeners = this.listeners.get(name);
+    listeners?.delete(listener);
+    if (listeners?.size === 0) {
+      this.listeners.delete(name);
+    }
+  }
+
+  notify(name: string) {
+    this.listeners.get(name)?.forEach((listener) => listener());
+  }
+
+  notifyAll() {
+    queueMicrotask(() => {
+      this.listeners.forEach((listeners) =>
+        listeners.forEach((listener) => listener()),
+      );
+    });
+  }
+
+  get(name: string) {
+    return this.state[name];
+  }
+
+  set(name: string, value: unknown) {
+    if (typeof value === "undefined") {
+      delete this.state[name];
+    } else {
+      this.state[name] = value;
+    }
+    this.notify(name);
   }
 
   load(locationKey: string) {
