+++ src/lib/locationState/stores/url/url-store.ts
@@ -1,20 +1,62 @@
-import { Store } from "../store";
+import { Listener, Store } from "../store";
 
-export class UrlStore extends Store {
+export class UrlStore implements Store {
   private readonly key: string;
+  listeners: Map<string, Set<Listener>> = new Map();
+  state: Record<string, unknown> = {};
 
   constructor({ key }: { key: string }) {
-    super();
     this.key = key;
   }
 
-  onSet() {
+  subscribe(name: string, listener: () => void) {
+    const listeners = this.listeners.get(name);
+    if (listeners) {
+      listeners.add(listener);
+    } else {
+      this.listeners.set(name, new Set([listener]));
+    }
+    return () => this.unsubscribe(name, listener);
+  }
+
+  private unsubscribe(name: string, listener: Listener) {
+    const listeners = this.listeners.get(name);
+    listeners?.delete(listener);
+    if (listeners?.size === 0) {
+      this.listeners.delete(name);
+    }
+  }
+
+  notify(name: string) {
+    this.listeners.get(name)?.forEach((listener) => listener());
+  }
+
+  notifyAll() {
+    queueMicrotask(() => {
+      this.listeners.forEach((listeners) =>
+        listeners.forEach((listener) => listener()),
+      );
+    });
+  }
+
+  get(name: string) {
+    return this.state[name];
+  }
+
+  set(name: string, value: unknown) {
+    if (typeof value === "undefined") {
+      delete this.state[name];
+    } else {
+      this.state[name] = value;
+    }
     // save in url
     const params = new URLSearchParams();
     params.set(this.key, JSON.stringify(this.state));
     const newSearch = params.toString();
     const newUrl = `${location.pathname}?${newSearch}`;
     history.replaceState(history.state, "", newUrl);
+
+    this.notify(name);
   }
 
   load() {
